Moved escape markdown func to telegram post file
